[REFACTOR] - pagina de inicio
