+++ harvester-app/src/framework/utils/scripts/interfaz/barraLateral.js
@@ -1,89 +1,53 @@
-/* eslint-disable no-unused-vars */
+const { ipcRenderer } = require('electron');
 
-/**
- * Carga dinámicamente el contenido HTML de un módulo en el contenedor principal.
- *
- * @param {string} seccion - Identificador de la sección a cargar
- *   (p.ej., 'inicio', 'analisis', 'plantillas', 'formulas', 'envios', 'usuario', 'gestionUsuarios').
- * @returns {void}
- */
-function cargarModulo(seccion) {
-  const rutasModulo = {
-    inicio:          '../vistas/moduloInicio.html',
-    analisis:        '../vistas/moduloAnalisis.html',
-    plantillas:      '../vistas/moduloPlantillas.html',
-    formulas:        '../vistas/moduloFormulas.html',
-    envios:          '../vistas/moduloEnvios.html',
-    usuario:         '../vistas/moduloUsuario.html',
-    gestionUsuarios: '../vistas/moduloGestionUsuarios.html',
-    tractores:  '../vistas/moduloTractores.html'
-  };
-
-  const contenedorVentanaPrincipal = document.querySelector('.ventana-principal');
-
-  if (contenedorVentanaPrincipal && rutasModulo[seccion]) {
-    // Guardar la vista actual una la pila antes de cambiar
-    const pila = JSON.parse(localStorage.getItem('vistaPila') || '[]');
-    const vistaActual = localStorage.getItem('seccion-activa');
-    if (vistaActual) {
-      // Agregar la vista actual en la pila y guardarla en localStorage
-      pila.push(vistaActual);
-      localStorage.setItem('vistaPila', JSON.stringify(pila));
-    }
-    fetch(rutasModulo[seccion])
-      .then(respuesta => respuesta.text())
-      .then(html => {
-        contenedorVentanaPrincipal.innerHTML = html;
-        actualizarBarraSuperior(seccion);
-        
-        // Inicializar el módulo según la sección
-        if (seccion === 'inicio') {
-          if (window.botonCargar)    window.botonCargar();
-          if (window.botonTractores)  window.botonTractores();
-          if (window.botonBorrar)    window.botonBorrar();
-        } else if (seccion === 'analisis') {
-            if (window.cargarDatosExcel)           window.cargarDatosExcel();
-            if (window.inicializarModuloAnalisis)  window.inicializarModuloAnalisis();
-        } else if (seccion === 'plantillas') {
-            if (window.inicializarModuloPlantillas) window.inicializarModuloPlantillas();
-        } else if (seccion === 'usuario') {
-            if (window.inicializarModuloUsuario)    window.inicializarModuloUsuario();
-        } else if (seccion ==='formulas'){
-          if (window.inicializarModuloFormulas) window.inicializarModuloFormulas();
-        } else if (seccion === 'tractores') {
-          if (window.inicializarModuloTractores) window.inicializarModuloTractores();
-        // Añadir más inicializaciones para otros módulos según sea necesario
-        }
-      })
-      .catch(error => {
-        console.error('Error al cargar el módulo:', error);
-      });
-  }
-}
-
-/**
- * Mapa de configuración para la barra superior: títulos e íconos según sección.
- *
- * @constant {Object.<string, { titulo: string, icono: string }>}
- */
-const infoBarraSuperior = {
-  inicio:          { titulo: 'Inicio',     icono: '../utils/iconos/Casa.svg' },
-  analisis:        { titulo: 'Análisis',   icono: '../utils/iconos/GraficaBarras.svg' },
-  plantillas:      { titulo: 'Plantillas', icono: '../utils/iconos/Portapapeles.svg' },
-  formulas:        { titulo: 'Fórmulas',   icono: '../utils/iconos/Funcion.svg' },
-  envios:          { titulo: 'Envíos',     icono: '../utils/iconos/Correo.svg' },
-  usuario:         { titulo: 'Usuario',    icono: '../utils/iconos/Usuario.svg' },
-  gestionUsuarios: { titulo: 'Usuario',    icono: '../utils/iconos/Usuario.svg' },
-  tractores:   { titulo: 'Tractores',    icono: '../utils/iconos/GraficaBarras.svg' }
+// Diccionario que incluye la ruta de la vista de cada módulo, el nombre de la sección, y el nombre del ícono de la sección.
+const informacionModulos = {
+  inicio:          [`${rutaBase}src/framework/vistas/paginas/inicio/inicio.ejs`, 'Inicio', 'Casa'],
+  tractores:       [`${rutaBase}src/framework/vistas/paginas/analisis/seleccionarTractor.ejs`, 'Perfil', 'Casa'],
+  analisis:        [`${rutaBase}src/framework/vistas/paginas/analisis/generarReporte.ejs`, 'Análisis', 'Casa'],
+  plantillas:      [`${rutaBase}src/framework/vistas/paginas/plantillas/listaPlantillas.ejs`, 'Plantillas', 'Casa'],
+  formulas:        [`${rutaBase}src/framework/vistas/paginas/formulas/listaFormulas.ejs`, 'Formulas', 'Casa'],
+  perfil:          [`${rutaBase}src/framework/vistas/paginas/usuarios/consultarPerfil.ejs`, 'Perfil', 'Casa'],
+  usuarios:        [`${rutaBase}src/framework/vistas/paginas/usuarios/listaUsuarios.ejs`, 'Perfil', 'Casa'],
 };
 
-/**
- * Inicializa la barra lateral: enlaza los botones de colapsar/expandir
- * y configura la navegación entre secciones.
- *
- * @returns {void}
- */
-function inicializarBarraLateral() {
+async function cargarModulo(modulo){
+  try {
+        var vista = await ipcRenderer.invoke('precargar-ejs', informacionModulos[modulo][0], { Seccion: informacionModulos[modulo][1], Icono : informacionModulos[modulo][2]});
+        window.location.href = vista;
+        localStorage.setItem('seccion-activa', modulo);
+    } catch (err) {
+        console.error('Error al cargar vista:', err);
+    }
+}
+
+configurarBotonesLaterales()
+function configurarBotonesLaterales(){
+
+  const botonesBarraLateral = document.querySelectorAll('.boton-sidebar');
+  botonesBarraLateral.forEach(boton => {
+    boton.addEventListener('click', () => {
+      const modulo = boton.getAttribute('data-seccion');
+      if (!modulo) return;
+
+      // Determinar cuál botón mostrar como activo visualmente
+      const seccionVisual = modulo === 'gestionUsuarios' ? 'usuario' : modulo;
+
+      // Desactivar todos los botones
+      botonesBarraLateral.forEach(boton => boton.classList.remove('activo'));
+      cargarModulo(modulo);
+
+      // Activar el botón correspondiente
+      document
+        .querySelectorAll(`.boton-sidebar[data-seccion='${seccionVisual}']`)
+        .forEach(botonItem => botonItem.classList.add('activo'));
+    })
+  });
+}
+
+cerrarBarraLateral()
+function cerrarBarraLateral() {
+  // const barraLateralExpandida = document.getElementById('barraLateralExpandida');
   const botonColapsar         = document.getElementById('botonColapsar');
   const botonExpandir         = document.getElementById('botonExpandir');
   const barraLateralExpandida = document.getElementById('barraLateralExpandida');
@@ -106,46 +70,6 @@ function inicializarBarraLateral() {
       barraLateralExpandida.style.display = 'flex';
     });
   }
-
-  activarBotonesBarraLateral();
-  
-  // Siempre iniciar en el módulo 'inicio'
-  localStorage.setItem('seccion-activa', 'inicio');
-  aplicarActivoDesdeAlmacenamiento();
-}
-
-/**
- * Añade listeners de navegación a los botones de la barra lateral.
- *
- * @returns {void}
- */
-function activarBotonesBarraLateral() {
-  const botonesSidebar = document.querySelectorAll('.boton-sidebar');
-
-  botonesSidebar.forEach(boton => {
-    boton.addEventListener('click', () => {
-      const seccion = boton.getAttribute('data-seccion');
-      if (!seccion || seccion === 'tema') return;
-
-      // Guardamos la sección real
-      localStorage.setItem('seccion-activa', seccion);
-
-      // Determinar cuál botón mostrar como activo visualmente
-      const seccionVisual = seccion === 'gestionUsuarios' ? 'usuario' : seccion;
-
-      // Desactivar todos los botones
-      botonesSidebar.forEach(botonItem => botonItem.classList.remove('activo'));
-
-      // Activar el botón correspondiente
-      document
-        .querySelectorAll(`.boton-sidebar[data-seccion='${seccionVisual}']`)
-        .forEach(botonItem => botonItem.classList.add('activo'));
-
-      // Actualizar barra superior y cargar contenido
-      actualizarBarraSuperior(seccion);
-      cargarModulo(seccion);
-    });
-  });
 }
 
 /**
@@ -166,51 +90,6 @@ function aplicarActivoDesdeAlmacenamiento() {
     .querySelectorAll(`.boton-sidebar[data-seccion='${seccionVisual}']`)
     .forEach(botonItem => botonItem.classList.add('activo'));
 
-  actualizarBarraSuperior(seccion);
-  cargarModulo(seccion);
 }
 
-/**
- * Actualiza el título, el ícono y el estado del botón de regresar
- * de la barra superior según la sección actual.
- *
- * @param {string} seccion - Identificador de la sección activa.
- * @returns {void}
- */
-function actualizarBarraSuperior(seccion) {
-  const elementoTituloBarraSuperior = document.getElementById('tituloBarraSuperior');
-  const elementoIconoBarraSuperior  = document.getElementById('iconoBarraSuperior');
-  const botonRegresar               = document.getElementById('botonRegresar');
-
-  if (!elementoTituloBarraSuperior || !elementoIconoBarraSuperior || !infoBarraSuperior[seccion]) {
-    return;
-  }
-
-  // Cambiar título e ícono
-  elementoTituloBarraSuperior.textContent = infoBarraSuperior[seccion].titulo;
-  elementoIconoBarraSuperior.src         = infoBarraSuperior[seccion].icono;
-
-  // Ocultar o mostrar botón de regresar
-  const seccionesSinRegresar = ['inicio', 'envios', 'plantillas', 'usuario','formulas'];
-  if (seccionesSinRegresar.includes(seccion)) {
-    botonRegresar.style.display                  = 'none';
-    elementoTituloBarraSuperior.style.marginLeft = '0px';
-  } else {
-    botonRegresar.style.display                  = 'flex';
-    elementoTituloBarraSuperior.style.marginLeft = '10px';
-    botonRegresar.onclick = () => {
-      const pila = JSON.parse(localStorage.getItem('vistaPila') || '[]');
-      if (pila.length > 0) {
-        // Sacar la última vista, guardar la pila actualizada y estableces la vista anterior como activa
-        const vistaAnterior = pila.pop();
-        localStorage.setItem('vistaPila', JSON.stringify(pila));
-        localStorage.setItem('seccion-activa', vistaAnterior);
-        cargarModulo(vistaAnterior);
-      }
-    };    
-  }
-}
-
-// Exponer la función de actualización de barra superior globalmente
-window.actualizarBarraSuperior = actualizarBarraSuperior;
-window.cargarModulo          = cargarModulo;
+aplicarActivoDesdeAlmacenamiento()
