--- harvester-app/src/framework/vistas/ModuloInicio.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="../utils/css/moduloInicio.css" rel="stylesheet" />
-</head>
-<body>
-  <div class="frame-principal-inicio">
-      <div class="frame-excel">
-          <div class="usar-excel">
-            <div class="texto-usar-excel">Usar Excel</div>
-            <img class="FolderPlusRojo" src="../utils/iconos/FolderMasRojo.svg" />
-            <div class="nombre-archivo">
-              <div class="texto-archivo">Sin Archivo Seleccionado</div>
-                <button id="boton-borrar" class="boton-borrar">
-                  <img class="borrar" src="../utils/iconos/X.svg" />
-                </button>
-            </div>
-            <label class="cargar">
-              <input type="file" accept=".xlsx" class="cargar-excel" />
-              <div class="texto-cargar">Cargar Archivo</div>
-            </label>
-            <button disabled class="avanzar-analisis">
-              <div class="texto-avanzar">Analizar documento</div>
-              <img class="arrow-back" src="../utils/iconos/FlechaAtras.svg" />
-            </button>
-          </div>
-      </div>
-    </div>
-</body>
-</html>
