+++ harvester-app/src/framework/vistas/includes/encabezado.ejs
@@ -13,9 +13,7 @@
   </div>
   <div class='formulas-brillante'>
     <div class='frame-10'>
-      <div class='sidebar-container'>
         <div id='contenedorBarraLateral'> <%- include('barraLateral.ejs') %> </div>
-      </div>
 
       <div class='main-container'>
           <div id='contenedorBarraSuperior'> <%- include('barraSuperior.ejs') %> </div>
