+++ harvester-app/src/framework/vistas/includes/piePagina.ejs
@@ -6,6 +6,6 @@
   <script>
       const rutaBase = <%- JSON.stringify(rutaBase) %>;
   </script>
-  <script src='<%= rutaBase %>src/framework/utils/scripts/interfaz/barraLateral2.js'></script>
+  <script src='<%= rutaBase %>src/framework/utils/scripts/interfaz/barraLateral.js'></script>
 </body>
 </html>
